Move wd tests to subdirectory.
