+++ examples/mocha/wd/setup-per-test-client.js
@@ -1,5 +1,5 @@
 // Overwite adapter with "per test client" version.
-var Adapter = require("../../index").adapters.mocha;
+var Adapter = require("../../../index").adapters.mocha;
 global.adapter = new Adapter(null, {
   client: {
     perTest: true
