+++ examples/mocha/wd/test/notes.spec.js
@@ -5,7 +5,7 @@ var adapter = global.adapter;
 
 // TODO: Needs lazy require for `wd`.
 var asserters = require("wd").asserters;
-var rowdy = require("../../../index");
+var rowdy = require("../../../../index");
 var jsFn = rowdy.helpers.js.fn;
 
 describe("notes", function () {
