+++ package.json
@@ -41,10 +41,11 @@
     "lint": "eslint .",
     "check": "npm run lint",
     "check-all": "npm run lint && npm run test",
-    "test": "npm run test-default-config && npm run test-manual-config && npm run test-per-test-client",
-    "test-default-config": "cd examples/mocha && mocha",
-    "test-manual-config": "cd examples/mocha && mocha --require ./setup-manual-config",
-    "test-per-test-client": "cd examples/mocha && mocha --require ./setup-per-test-client",
+    "test": "npm run test-wd",
+    "test-wd": "npm run test-wd-default-config && npm run test-wd-manual-config && npm run test-wd-per-test-client",
+    "test-wd-default-config": "cd examples/mocha/wd && mocha",
+    "test-wd-manual-config": "cd examples/mocha/wd && mocha --require ./setup-manual-config",
+    "test-wd-per-test-client": "cd examples/mocha/wd && mocha --require ./setup-per-test-client",
     "install-selenium": "selenium-standalone install"
   }
 }
